Refactoring
Divided the code into multiple classes for improved readability and maintainability.
