+++ Assets/RealCanvas/Scripts/CanvasRecognitionDLL.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f2dbc1b84ad1c0343a1c62bd19431449
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
