+++ Assets/RealCanvas/Scripts/MaterialController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 09ee4920d17ebf3439c65999fd8da838
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
